+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.11.7
+- fix: Use rt import path when using beetle bitswap feature.
+
 # 0.11.6
 - feat: Add RepoInsertPin::provider and RepoInsertPin::providers. [PR 180](https://github.com/dariusc93/rust-ipfs/pull/180)
 - chore: Add RelayConfig::unbounded. [PR 181](https://github.com/dariusc93/rust-ipfs/pull/181)